Module legion_systems::prelude

Re-exports

pub use crate::resource::ResourceSet;

pub use crate::resource::Resources;

pub use crate::schedule::Executor;

pub use crate::schedule::Runnable;

pub use crate::schedule::Schedulable;

pub use crate::schedule::Schedule;

Structs

BitSet

System
The concrete type which contains the system closure provided by the user. This struct should not be instantiated directly, and instead should be created using SystemBuilder.

SystemBuilder
The core builder of System types, which are systems within Legion. Systems are implemented as singular closures for a given system - providing queries which should be cached for that system, as well as resource access and other metadata.